Complete tab functionality for part 2
I finished part 2 of the assignment. It was pretty tricky. The logic behind using classes is a bit cumbersome to follow at times. The biggest difficulty I found was actually the naming conventions in the HTML. There were such slight variations in names that it was easy to use the wrong one and not notice.