Complete? the first stretch task
I really didn't understand exactly what the assignment was asking. Based on my interpretation of what was being asked, I was able to put together something that was mostly effective. Making a component that created all instances of TabLink wasn't too bad, but dynamically storing the selected tab was challenging.